Rough implementation of continuous polling from the sim
Each worker launches a background thread which checks for new timesteps
from the simulation and builds a new pkd tree on this data. We then
toggle the active pkd and notify the master (back in sg::InSituSpheres)
that we're dirty and should be re-committed to update. When we re-commit
we pick data from the active pkd, which is the one we've just built with
the new timestep.

While this goes on we're looking for a new timestep to put in the other
one and ping-ponging between these two pkds. The master node is notified
of a changed by the first worker sending them the world bounds again and
the master similarly has a polling thread running and watching for the
bounds to be sent at which point it marks us dirty and we're
recommitted.