Fix bug on not picking more than one box, reveals bug assigning boxes
When there are more boxes than workers it is possible (and in fact very
likely) that the one boxes are given to workers will result in
non-convex non-disjoint regions being rendered resulting in compositing
artifacts. See for example a 4^3 render grid with 8 workers (+1
display), the regions assigned intersect each other and we have issues.

We need a way to assign convex, disjoint subregions of the grid to
workers, striding with modulo will not work.